Extract collection method tck in a separate interface + mark the potential duplicate methods for removal
